+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 ### Fixed and Changed
 
 * Fixing minor issue in setting the parent of qml objects.
-
+* Fixing issue in GUI library in stacking the Timeline after each Video Call request.
 
 
 ## [0.0.24]