+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 ### Fixed and Changed
 
 * Fixing minor issue in setting the parent of qml objects.
+* Addding Server Discovery
+* Integrate Cibalogin and Password login
+* Adding login option combo box