+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
+## [0.0.26]
+### Fixed
+
+* Fixing build error issue in applications which integrated with library.
+
+
 ## [0.0.25]
 ### Fixed and Changed